refactor: StatusChip 공통 메서드 추출
